refactor: int인 id를 Long으로 타입변경.

+++ src/main/java/io/cavia/trader/module/member/repository/MybatisMemberRepository.java
@@ -24,7 +24,7 @@ public class MybatisMemberRepository implements MemberRepository {
     }
 
     @Override
-    public int updateNickname(int id, String nickname, LocalDateTime nicknameUpdatedAt) {return memberMapper.updateNickname(id,nickname,nicknameUpdatedAt);}
+    public int updateNickname(Long id, String nickname, LocalDateTime nicknameUpdatedAt) {return memberMapper.updateNickname(id,nickname,nicknameUpdatedAt);}
 
     @Override
     public int updatePassword(int id, String password, LocalDateTime passwordUpdatedAt) {return memberMapper.updatePassword(id,password,passwordUpdatedAt);}
